Add and implement Meta trait
